+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- **New:** Added literal type information to fields to provide additional information about the field type. (#51)
+
 ### Changed
 - Initial development WIP - unpublished